Moved pages into components and created folders for each page
