+++ ecosphere-goods/src/components/pages/user-credentials-page/user-credentials/RegisterPanel.js
