--- ecosphere-goods/src/pages/HomePage.js
@@ -1,16 +1,0 @@
-import React from 'react'
-import Hero from '../components/Hero'
-import SpecialItems from '../components/SpecialItems'
-import ClimateFacts from '../components/ClimateFacts'
-
-function HomePage() {
-  return (
-    <div>
-      <Hero/>
-      <SpecialItems/>
-      <ClimateFacts/>
-    </div>
-  )
-}
-
-export default HomePage
